Some refactoring to make it easier to send to api
